PoC_TreinamentoDeYolo/requirements.txt
pyautogui
keyboard
pillow
opencv-python
numpy